+++ bit/cli/mutate_seqs.py
@@ -11,7 +11,6 @@ def build_parser():
     desc = """
         This script will mutate all sequences of a nucleotide or amino-acid multifasta with
         the specified mutation rate. By default it only swaps bases, but it can optionally introduce indels also.
-        Use `-v` or `--version` flag for version info.
         """
 
     parser = argparse.ArgumentParser(